fix(output): Treat write() parameters correctly.
The cause of this is because the line:

  process.stdout.write = function () {
    output += Array.prototype.join.call(arguments, '');
  };

is not actually correct. in fact, the arguments to write are (data[, encoding][, callback]), so joining them doesn't make sense. Evidently in 8.x node started passing a callback function in.

Fix for GitHub issue #1: With Node 8.1.0 get weird function output when using. #1